move lockmap to util package
